注文管理画面の新規注文,調理中,呼び出し中の各セクションカードをコンポーネント化
