+++ src/types/order.ts
@@ -1,0 +1,7 @@
+export interface Order {
+  id: number;
+  orderNumber: number;
+  quantity: number;
+  status: "pending" | "ready" | "completed";
+  createdAt: string;
+}
